+++ DataDefinitions/Ship.cs
@@ -275,11 +275,9 @@ namespace EddiDataDefinitions
         public List<Hardpoint> hardpoints { get; set; }
         public List<Compartment> compartments { get; set; }
         public List<LaunchBay> launchbays { get; set; }
-
+        public decimal activeFuelReservoirCapacity { get; set; }
         public string paintjob { get; set; }
 
-        public decimal? fuelremaining { get; set; } // Current fuel level
-        public decimal? fuelpercent => percentFuel();
         public decimal? fueltankcapacity { get; set; } // Core capacity
         public decimal? fueltanktotalcapacity { get; set; } // Capacity including additional tanks
         public decimal maxfuel { get; set; }
@@ -301,7 +299,7 @@ namespace EddiDataDefinitions
             launchbays = new List<LaunchBay>();
         }
 
-        public Ship(long EDID, string EDName, string Manufacturer, List<Translation> PhoneticManufacturer, string Model, List<Translation> PhoneticModel, string Size, int? MilitarySize)
+        public Ship(long EDID, string EDName, string Manufacturer, List<Translation> PhoneticManufacturer, string Model, List<Translation> PhoneticModel, string Size, int? MilitarySize, decimal reservoirFuelTankSize)
         {
             this.EDID = EDID;
             this.EDName = EDName;
@@ -315,6 +313,7 @@ namespace EddiDataDefinitions
             hardpoints = new List<Hardpoint>();
             compartments = new List<Compartment>();
             launchbays = new List<LaunchBay>();
+            this.activeFuelReservoirCapacity = reservoirFuelTankSize;
         }
 
         public override string ToString()
@@ -445,6 +444,7 @@ namespace EddiDataDefinitions
                 phoneticmodel = template.phoneticmodel;
                 size = template.size;
                 militarysize = template.militarysize;
+                activeFuelReservoirCapacity = template.activeFuelReservoirCapacity;
                 if (Role == null)
                 {
                     Role = EddiDataDefinitions.Role.MultiPurpose;
@@ -457,15 +457,6 @@ namespace EddiDataDefinitions
         public void NotifyPropertyChanged(string propName)
         {
             PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propName));
-        }
-
-        private decimal? percentFuel()
-        {
-            if (fueltanktotalcapacity != null && fuelremaining != null && fueltanktotalcapacity > 0)
-            {
-                return Math.Round((decimal)(fuelremaining / fueltanktotalcapacity * 100), 1);
-            }
-            return null;
         }
     }
 }